feat: mmoved form to separate component and created form with input commponent
